Added Dark Ending functionality
Added this via an option for boss split that makes it split on the map when the boss is unlocking, rather than when entering the boss. It would be preferable to do without this option and have it always split on unlock, but this would have an adverse affect on legacy splits that split on boss entrance. I felt this method was preferable to adding an entire second boss split list for unlock.