Fixed buggy IL split when getting forced out to map
`playing` has variable behavior depending on how you exit the level (double spacebar press / A / S). It is more consistent to use old.playing, since it still gives us the info that the level was beaten, but is unaffected by your exit method. However, it seems frame perfect A/S exits prevent this address from being updated, so checking whether the IL time has been writen is a more consistent method. That said, IL time gets erased when entering a replay, so old.playing is still used as a backup in case replay is entered.